# webapp/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
      "Forecast": {
        "additionalProperties": false,
        "properties": {
          "deterministic_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Run"
              },
              {
                "type": "null"
              }
            ]
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "pictograms": {
            "additionalProperties": {
              "$ref": "#/components/schemas/PictogramSeries"
            },
            "default": {},
            "title": "Pictograms",
            "type": "object"
          },
          "run": {
            "$ref": "#/components/schemas/Run"
          },
          "step_hours": {
            "minimum": 1.0,
            "title": "Step Hours",
            "type": "integer"
          },
          "steps": {
            "items": {
              "format": "date-time",
              "type": "string"
            },
            "title": "Steps",
            "type": "array"
          },
          "variables": {
            "additionalProperties": {
              "$ref": "#/components/schemas/VariableSeries"
            },
            "title": "Variables",
            "type": "object"
          }
        },
        "required": [
          "location",
          "run",
          "steps",
          "step_hours",
          "variables"
        ],
        "title": "Forecast",
        "type": "object"
      },
      "GeocodeOut": {
        "properties": {
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/Place"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "query",
          "results"
        ],
        "title": "GeocodeOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Location": {
        "additionalProperties": false,
        "properties": {
          "elevation_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elevation M"
          },
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon",
            "type": "number"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "required": [
          "lat",
          "lon"
        ],
        "title": "Location",
        "type": "object"
      },
...
      "PictogramItem": {
        "additionalProperties": false,
        "properties": {
          "class": {
            "title": "Class",
            "type": "string"
          },
          "level": {
            "minimum": 1.0,
            "title": "Level",
            "type": "integer"
          },
          "pictogram": {
            "title": "Pictogram",
            "type": "string"
          }
        },
        "required": [
          "pictogram",
          "level",
          "class"
        ],
        "title": "PictogramItem",
        "type": "object"
      },
      "PictogramSeries": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PictogramItem"
            },
            "title": "Items",
            "type": "array"
          },
          "scheme": {
            "title": "Scheme",
            "type": "string"
          }
        },
        "required": [
          "scheme",
          "items"
        ],
        "title": "PictogramSeries",
        "type": "object"
      },
      "Place": {
        "properties": {
          "admin1": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin1"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "elevation_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elevation M"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "required": [
          "name",
          "lat",
          "lon"
        ],
        "title": "Place",
        "type": "object"
      },
...
      "Run": {
        "additionalProperties": false,
        "properties": {
          "init_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Init Time"
          },
          "members": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Members"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "source"
        ],
        "title": "Run",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VariableSeries": {
        "additionalProperties": false,
        "properties": {
          "deterministic": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deterministic"
          },
          "kind": {
            "enum": [
              "instant",
              "sum"
            ],
            "title": "Kind",
            "type": "string"
          },
          "quantiles": {
            "additionalProperties": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "title": "Quantiles",
            "type": "object"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "window_hours": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Hours"
          }
        },
        "required": [
          "unit",
          "kind",
          "quantiles"
        ],
        "title": "VariableSeries",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Probabilistic forecasts as quantiles plus VSUP pictograms. Read-only.",
    "title": "VSUP meteogram API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/forecast": {
      "get": {
        "description": "The whole forecast the source has - the client picks the days it shows -\nas quantiles per step plus the product's pictograms.",
        "operationId": "forecast_api_forecast_get",
        "parameters": [
          {
            "description": "Rounded to 2 decimals.",
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "description": "Rounded to 2 decimals.",
              "maximum": 90,
              "minimum": -90,
              "title": "Lat",
              "type": "number"
            }
          },
          {
            "description": "Rounded to 2 decimals.",
            "in": "query",
            "name": "lon",
            "required": true,
            "schema": {
              "description": "Rounded to 2 decimals.",
              "maximum": 180,
              "minimum": -180,
              "title": "Lon",
              "type": "number"
            }
          },
          {
            "description": "A product id from /api/products; default: the first.",
            "in": "query",
            "name": "product",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A product id from /api/products; default: the first.",
              "title": "Product"
            }
          },
          {
            "in": "query",
            "name": "variant",
            "required": false,
            "schema": {
              "default": "ensemble",
              "enum": [
                "ensemble",
                "hres"
              ],
              "title": "Variant",
              "type": "string"
            }
          },
          {
            "description": "Put into location.name as given.",
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Put into location.name as given.",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forecast"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The source has no data for this place."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The product lacks this variant."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The upstream source failed or sent unusable data."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The upstream source did not answer in time."
          }
        },
        "summary": "Forecast",
        "tags": [
          "forecast"
        ]
      }
    },
    "/api/geocode": {
      "get": {
        "description": "Places matching `q`, best first; an empty list rather than an error\nwhen nothing matches. Fine to call while the user types.",
        "operationId": "geocode_api_geocode_get",
        "parameters": [
          {
            "description": "Place name; fewer than 2 characters find nothing.",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "Place name; fewer than 2 characters find nothing.",
              "maxLength": 100,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "en",
              "enum": [
                "de",
                "en"
              ],
              "title": "Lang",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "count",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 20,
              "minimum": 1,
              "title": "Count",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeocodeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The source has no data for this place."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The upstream source failed or sent unusable data."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The upstream source did not answer in time."
          }
        },
        "summary": "Geocode",
        "tags": [
          "forecast"
        ]
      }
    },
...
  }
}